add board_detail
