+++ Board/.settings/org.eclipse.core.resources.prefs
@@ -1,2 +1,6 @@
 eclipse.preferences.version=1
 encoding/Board_main.py=utf-8
+<<<<<<< HEAD
+=======
+encoding/config.py=utf-8
+>>>>>>> 8df0859ab6929413e4d4b4c26c75e0bf2a10095c
