+++ Board/Board_main.py
@@ -9,8 +9,13 @@ from os.path import dirname, join
 from sqlalchemy import create_engine, Column, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import scoped_session, sessionmaker
-
+from sqlalchemy.types import DateTime, Boolean
+from sqlalchemy.sql import functions
+from datetime import datetime,date,time
 from config import *
+
+import hashlib
+import urllib
 
 
 app = Flask(__name__)
@@ -21,12 +26,69 @@ oid = OpenID(app, join(dirname(__file__), 'openid_store'))
 app.config.update(
         SQLALCHEMY_DATABASE_URI = 'postgres://uvkxbyzicejuyd:FzhZqstwa1YQ7FVPNAId0GO_4l@ec2-54-197-241-91.compute-1.amazonaws.com:5432/d22mrqavab61bp',
         #'postgresql://postgres:1111@localhost:5432/Board',
+        DATABASE_URI = 'postgresql://postgres:1234@localhost:5432/pos',
         SECRET_KEY = 'development key',
         DEBUG = True
     )
+
+db = SQLAlchemy(app)
+app.config.from_envvar('FLASKR_SETTINGS', silent=True)
+
+#db settings
+
+engine = create_engine(app.config['SQLALCHEMY_DATABASE_URI'], convert_unicode=True)
+db_session =scoped_session(sessionmaker(autocommit=False,
+                                         autoflush=False,
+                                         bind=engine))
+Base = declarative_base()
+Base.query = db_session.query_property()
+
+class Comment(db.Model):
+    __tablename__ = 'comments'
+    id = db.Column(db.Integer, primary_key=True)
+    email = db.Column(db.String(200))
+    comment = db.Column(db.String(500))
+
+    def __init__(self, email, comment):
+        self.email = email
+        self.comment = comment
+
+
+class Post(db.Model):
+    __tablename__='posts'
+    
+    board_id = db.Column(Integer, primary_key=True)
+    category = db.Column(db.String(10))
+    subject = db.Column(db.String(50))
+    status = db.Column(db.String(20))
+    contents = db.Column(db.String(500))
+    author_id= db.Column(db.String(20))
+    created_at= db.Column(DateTime(timezone=True), nullable=False,
+                                     default=functions.now())
+
+    def __init__(self,category,subject,status,contents):
+        self.category = category
+        self.subject = subject
+        self.status = status
+        self.contents = contents
+        
+    def __repr__(self):
+        return '<Post %s,%s,%s,%s>' % self.category, self.subject,\
+        self.status, self.contents
 # def get_user():
 #    return g.db.get('oid-' + session.get('openid', ''))
 
+class User(Base):
+    __tablename__ = 'users'
+    id = Column(Integer, primary_key=True)
+    name = Column(String(60))
+    email = Column(String(200))
+    admin = Column(Boolean(create_constraint=True, name=None))
+    
+    def __init__(self, name, email):
+        self.name = name
+        self.email = email
+        
 def init_db():    
     db.create_all()
     Base.metadata.create_all(bind=engine) 
@@ -35,15 +97,17 @@ def init_db():
 def index():      
     return render_template('index.html')
 
-@app.route('/top')
+@app.route('/board_top')
 def top():
-    return render_template('top.html')  
+    return render_template('board_top.html')
 
+#글리스트 부분
 @app.route('/board_list')
 def board_list():
     post_list = Post.query.all()
     return render_template('board_list.html', post_list=post_list)
 
+#글쓰기부분
 @app.route('/board_insert', methods=['GET','POST'])
 def board_insert(): 
     if request.method == 'POST':  
@@ -58,6 +122,11 @@ def board_insert():
         
         return redirect(url_for('board_list'))
     return render_template('board_insert.html')
+
+@app.route('/board_detail')
+def board_detail():
+    post_detail = Post.query.all()
+    return render_template('board_detail.html', post_detail=post_detail)
 
 @app.route('/editor')
 def editor():
@@ -107,11 +176,6 @@ def set_img(resp):
     gravatar_url += urllib.urlencode( {'d': 'mm' ,'s': str(size)} )     
     return gravatar_url, email_gra     
       
-      
-@app.route('/post')
-def post():
-    return render_template('example.html')
-
 @app.route('/logout')
 def logout():
     session.pop('id', None)
@@ -123,7 +187,6 @@ def contents():
     comm_list = Comment.query.all()
     return render_template('contents.html', comm_list=comm_list)
     
-
 if __name__ == '__main__':
     init_db()
     app.run(debug=True, host='0.0.0.0', port=int(environ.get('PORT',5000)))
