+++ Board/static/css/style.css
@@ -1,0 +1,69 @@
+.centered{ 
+	padding-top:20%;   
+	padding-right:100px;
+	padding-left:100px;           
+}        
+   .cursor{
+                   cursor:pointer;
+   }
+           P {
+                font-size: x-large;        
+                }  
+    .box {
+         margin: 10px;
+         padding: 15px;
+         height: 15px;
+         width: 300px;
+         text-align: center;
+         border: 1px solid ;
+         font-size: 10pt; 
+         cursor:pointer;
+         
+        }
+   .pTitle{
+              font-size:20pt;
+               font-weight:bolder;
+               text-align: center;
+        }
+   .lFont{
+               font-size: 10pt;
+               text-align: right;
+       }
+        .nav-list {
+                position:fixed;
+                top: 35%;
+                
+                width: 21.7%;
+        }
+       .go-top{
+                        position: fixed;
+                        bottom:2em;
+                        left:2em;
+                        text-decoration: none;
+                        color:white;
+                        background-color:gray;
+                        font-size: 12px;
+                        padding:1em;
+                        
+                        }
+           .go-top:hover{
+                        
+                background-color:black;
+               
+                           }
+   .date{
+                   text-align:center;
+                   font-size: 12px;
+                                              
+           }
+          .body{
+                          padding-top:20%;
+                          padding-right:100px;
+                          padding-left:100px;
+                          
+          }
+
+table{
+	height:30px;
+	background-color:#428BCA;
+}
