+++ Board/templates/board_detail.html
@@ -1,13 +1,19 @@
-{% extends "top.html" %}
+{% extends "board_top.html" %}
 {% block body %}
 {{ super() }}
 <div class="container">
-<h1>제안서 상황</h1>
-<br/><br/>
-	<div class="row">
-		<div class="col-md-12">이 제안은 {{ g.detail.status }}입니다.</div>
-		<div class="col-md-12">{{ g.detail.contents }}</div>
-	</div>
+	<h1>제안서 상황</h1>
+	<br/><br/>
+	
+	<table class="table table-striped">
+		{% for post in post_detail %}
+		<thead> 이 제안은 {{ post.status }} 입니다.</thead>
+		<tbody>
+			<tr><td>{{ post.contents }}</td></tr>
+		</tbody>
+		{% endfor %}
+	</table>
+
 	
 </div>
