+++ Board/templates/board_top.html
@@ -21,7 +21,7 @@
                         <li class="active"><a href="#">제안서</a></li>
                         <li><a href="{{ url_for('board_list') }}">목록보기</a></li>
                         <li><a href="{{ url_for('board_insert') }}">제안쓰기</a></li>
-                  
+                  		<li><a href="{{ url_for('board_detail')}}">제안서 상황</a></li>
                         <li><a href="{{url_for('logout') }}">로그아웃</a></li>
         </div><!-- /.navbar-collapse -->
 </nav>
