+++ Board/templates/index.html
@@ -1,0 +1,28 @@
+{% extends "layout.html" %}
+{% block body %}
+<div class="centered container" style="padding:0 19px">
+        
+        <div class="row centered" >
+        
+                <div class="col-xs-6 col-md-4 ">
+                         <img src="static/images/dodo.png" />
+                </div>                        
+                <div class="col-xs-6 col-sm-12 col-md-6 col-md-offset-2">
+                         <h1> 도도 제품 제안서</h1>        
+                         <br><a class="box" href="{{url_for('login')}}"> 스포카 계정으로 로그인하기 </a>                                                    
+                </div>
+        </div>
+<div class="row body" style="padding: 0 19px">
+        <div class="col-xs-12 col-sm-12 col-md-12 col-centered"> 
+        <hr />
+        <p>
+                도도 제품 아이디어 제안서 서비스는 스포카 임직원분들이 자유롭게 아이디어를 올리고 다른 분들의 아이디어에 의견을 공유할 수 있는 서비스입니다. <br>
+                (주) 스포카 개발팀은 제안서의 아이디어를 첫번째 우선순위로 업데이트 계획을 잡습니다. 좋은 아이디어가 있다면 적극적으로 공유해주시면 서비스 업데이트 방향 설저에 큰 도움이 됩니다. 
+        </p>
+        </div>
+</div>
+</div>
+<script type="text/javascript">
+</script>
+{% endblock %}
+
